could probably be done more elegantly, but I think it's not bad
