+++ crates/sui-graphql-rpc/src/context_data/db_data_provider.rs
@@ -45,7 +45,7 @@ use crate::{
     },
 };
 use async_graphql::connection::{Connection, Edge};
-use diesel::{pg::Pg, ExpressionMethods, OptionalExtension, PgConnection, QueryDsl, RunQueryDsl};
+use diesel::{ExpressionMethods, OptionalExtension, PgConnection, QueryDsl, RunQueryDsl};
 use move_core_types::language_storage::StructTag;
 use std::str::FromStr;
 use sui_indexer::{
@@ -90,13 +90,10 @@ use sui_types::{
     Identifier,
 };
 
-use super::{
-    db_backend::{Explain, GenericQueryBuilder},
-    DEFAULT_PAGE_SIZE,
-};
+use super::{db_backend::GenericQueryBuilder, DEFAULT_PAGE_SIZE};
 
 #[cfg(feature = "pg_backend")]
-use super::pg_backend::QueryBuilder;
+use super::pg_backend::{PgQueryExecutor, QueryBuilder};
 
 #[derive(thiserror::Error, Debug, Eq, PartialEq)]
 pub enum DbValidationError {
@@ -122,23 +119,6 @@ pub enum DbValidationError {
     InvalidOwnerType,
     #[error("Query cost exceeded - cost: {0}, limit: {1}")]
     QueryCostExceeded(u64, u64),
-}
-
-pub fn extract_cost(explain_result: &str) -> Result<f64, Error> {
-    let parsed: serde_json::Value =
-        serde_json::from_str(explain_result).map_err(|e| Error::Internal(e.to_string()))?;
-    if let Some(cost) = parsed
-        .get(0)
-        .and_then(|entry| entry.get("Plan"))
-        .and_then(|plan| plan.get("Total Cost"))
-        .and_then(|cost| cost.as_f64())
-    {
-        Ok(cost)
-    } else {
-        Err(Error::Internal(
-            "Failed to get cost from query plan".to_string(),
-        ))
-    }
 }
 
 pub(crate) struct PgManager {
@@ -171,51 +151,49 @@ impl PgManager {
             .map_err(|e| Error::Internal(e.to_string()))
     }
 
-    /// Takes a query_builder_fn that returns Result<QueryFragment> and a lambda to execute the query
-    /// Spawns a blocking task that determines the cost of the query fragment
-    /// And if within limits, then executes the query
-    async fn run_query_async_with_cost<T, Q, QResult, EF, E, F>(
-        &self,
-        mut query_builder_fn: Q,
-        execute_fn: EF,
-    ) -> Result<T, Error>
-    where
-        Q: FnMut() -> Result<QResult, Error> + Send + 'static,
-        QResult: QueryDsl
-            + RunQueryDsl<Pg>
-            + diesel::query_builder::QueryFragment<diesel::pg::Pg>
-            + diesel::query_builder::Query
-            + diesel::query_builder::QueryId
-            + Send
-            + 'static,
-        EF: FnOnce(QResult) -> F + Send + 'static,
-        F: FnOnce(&mut PgConnection) -> Result<T, E> + Send + 'static,
-        E: From<diesel::result::Error> + std::error::Error + Send + 'static,
-        T: Send + 'static,
-    {
-        let max_db_query_cost = self.limits.max_db_query_cost;
-        self.inner
-            .spawn_blocking(move |this| {
-                let query = query_builder_fn()?;
-                let explain_result: String = this
-                    .run_query(|conn| query.explain().get_result(conn))
-                    .map_err(|e| Error::Internal(e.to_string()))?;
-                let cost = extract_cost(&explain_result)?;
-                if cost > max_db_query_cost as f64 {
-                    return Err(DbValidationError::QueryCostExceeded(
-                        cost as u64,
-                        max_db_query_cost,
-                    )
-                    .into());
-                }
+    // / Takes a query_builder_fn that returns Result<QueryFragment> and a lambda to execute the query
+    // / Spawns a blocking task that determines the cost of the query fragment
+    // / And if within limits, then executes the query
+    // async fn run_query_async_with_cost<T, Q, QResult, EF, E, F>(
+    //     &self,
+    //     mut query_builder_fn: Q,
+    //     execute_fn: EF,
+    // ) -> Result<T, Error>
+    // where
+    //     Q: FnMut() -> Result<QResult, Error> + Send + 'static,
+    //     QResult: diesel::query_builder::QueryFragment<diesel::pg::Pg>
+    //         + diesel::query_builder::Query
+    //         + diesel::query_builder::QueryId
+    //         + Send
+    //         + 'static,
+    //     EF: FnOnce(QResult) -> F + Send + 'static,
+    //     F: FnOnce(&mut PgConnection) -> Result<T, E> + Send + 'static,
+    //     E: From<diesel::result::Error> + std::error::Error + Send + 'static,
+    //     T: Send + 'static,
+    // {
+    //     let max_db_query_cost = self.limits.max_db_query_cost;
+    //     self.inner
+    //         .spawn_blocking(move |this| {
+    //             let query = query_builder_fn()?;
+    //             let explain_result: String = this
+    //                 .run_query(|conn| query.explain().get_result(conn))
+    //                 .map_err(|e| Error::Internal(e.to_string()))?;
+    //             let cost = extract_cost(&explain_result)?;
+    //             if cost > max_db_query_cost as f64 {
+    //                 return Err(DbValidationError::QueryCostExceeded(
+    //                     cost as u64,
+    //                     max_db_query_cost,
+    //                 )
+    //                 .into());
+    //             }
 
-                let query = query_builder_fn()?;
-                let execute_closure = execute_fn(query);
-                this.run_query(execute_closure)
-                    .map_err(|e| Error::Internal(e.to_string()))
-            })
-            .await
-    }
+    //             let query = query_builder_fn()?;
+    //             let execute_closure = execute_fn(query);
+    //             this.run_query(execute_closure)
+    //                 .map_err(|e| Error::Internal(e.to_string()))
+    //         })
+    //         .await
+    // }
 }
 
 /// Implement methods to query db and return StoredData
@@ -1826,52 +1804,5 @@ impl From<SuiAddress> for NativeSuiAddress {
 impl From<&SuiAddress> for NativeSuiAddress {
     fn from(a: &SuiAddress) -> Self {
         NativeSuiAddress::try_from(a.as_slice()).unwrap()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_invalid_json() {
-        let explain_result = "invalid json";
-        let result = extract_cost(explain_result);
-        assert!(matches!(result, Err(Error::Internal(_))));
-    }
-
-    #[test]
-    fn test_missing_entry_at_0() {
-        let explain_result = "[]";
-        let result = extract_cost(explain_result);
-        assert!(matches!(result, Err(Error::Internal(_))));
-    }
-
-    #[test]
-    fn test_missing_plan() {
-        let explain_result = r#"[{}]"#;
-        let result = extract_cost(explain_result);
-        assert!(matches!(result, Err(Error::Internal(_))));
-    }
-
-    #[test]
-    fn test_missing_total_cost() {
-        let explain_result = r#"[{"Plan": {}}]"#;
-        let result = extract_cost(explain_result);
-        assert!(matches!(result, Err(Error::Internal(_))));
-    }
-
-    #[test]
-    fn test_failure_on_conversion_to_f64() {
-        let explain_result = r#"[{"Plan": {"Total Cost": "string_instead_of_float"}}]"#;
-        let result = extract_cost(explain_result);
-        assert!(matches!(result, Err(Error::Internal(_))));
-    }
-
-    #[test]
-    fn test_happy_scenario() {
-        let explain_result = r#"[{"Plan": {"Total Cost": 1.0}}]"#;
-        let result = extract_cost(explain_result).unwrap();
-        assert_eq!(result, 1.0);
     }
 }
